+++ src/ts/Core/Models/ABGroup.ts
@@ -69,15 +69,26 @@ class ZyngaFilteredABTest extends FilteredABTest {
     }
 }
 
+class ZyngaFilteredReverseABTest extends ZyngaFilteredABTest {
+    constructor(...groups: AllowedGroups[]) {
+        super(...groups);
+    }
+
+    public isValid(group: ABGroup): boolean {
+        return !this.isFiltered() && !super.isValid(group);
+    }
+}
+
 // For Unit Testing
 export const FakeEnabledABTest = new ABTest(16, 17);
 export const FakeDisabledABTest = new DisabledABTest(16, 17);
 export const FakeZyngaFilteredABTest = new ZyngaFilteredABTest(16, 17);
 export const FakeReverseABTest = new ReverseABTest(16, 17);
+export const FakeZyngaFilteredReverseABTest = new ZyngaFilteredReverseABTest(16, 17);
 
 // Active AB Tests
 export const MediationCacheModeAllowedTest = new ReverseABTest(5);
-export const MabDisabledABTest = new ZyngaFilteredABTest(7);
+export const MabReverseABTest = new ZyngaFilteredReverseABTest(7);
 export const GooglePlayDetectionTest = new ABTest(8);
 export const PrivacySDKTest = new ABTest(16);
 export const LoadV5NoInvalidation = new ABTest(13);