+++ src/ts/Core/__mocks__/Core.ts
@@ -20,6 +20,7 @@ import { IosPreferencesApi } from 'Core/Native/iOS/__mocks__/IosPreferencesApi';
 import { MainBundleApi } from 'Core/Native/iOS/__mocks__/MainBundleApi';
 import { NotificationApi } from 'Core/Native/iOS/__mocks__/NotificationApi';
 import { UrlSchemeApi } from 'Core/Native/iOS/__mocks__/UrlSchemeApi';
+import { TrackingManagerApi } from 'Core/Native/iOS/__mocks__/TrackingManagerApi';
 import { CacheApi } from 'Core/Native/__mocks__/CacheApi';
 import { ClassDetectionApi } from 'Core/Native/__mocks__/ClassDetectionApi';
 import { ConnectivityApi } from 'Core/Native/__mocks__/ConnectivityApi';
@@ -62,7 +63,8 @@ export const Core = jest.fn(() => {
                 MainBundle: new MainBundleApi(),
                 Notification: new NotificationApi(),
                 Preferences: new IosPreferencesApi(),
-                UrlScheme: new UrlSchemeApi()
+                UrlScheme: new UrlSchemeApi(),
+                TrackingManager: new TrackingManagerApi()
             }
         },
         NativeBridge: new NativeBridge(),