+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Pending
+# Tue Jul 21 18:51:38 UTC 2020
 
 * [Chore] Fix MRAID Error Logging [#16439](https://github.com/Applifier/unity-ads-webview/pull/16439)
 * [Load V5] Update Allow List GameIDs [#16438](https://github.com/Applifier/unity-ads-webview/pull/16438)