+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Pending
+# Tue Jul 21 21:43:41 UTC 2020
 
 * [VAST Endcard] Lock Screen Orientation on Show [#16495](https://github.com/Applifier/unity-ads-webview/pull/16495)
 * [AUI] Make MAB reverse AB test explicit [#16436](https://github.com/Applifier/unity-ads-webview/pull/16436)