+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Pending
+
+* Add load and fill event manager [#16401](https://github.com/Applifier/unity-ads-webview/pull/16401)
+
 # Tue Jul 21 21:43:41 UTC 2020
 
 * [VAST Endcard] Lock Screen Orientation on Show [#16495](https://github.com/Applifier/unity-ads-webview/pull/16495)